+++ server/graphql/stock_line/src/lib.rs
@@ -2,96 +2,16 @@ pub mod mutations;
 use async_graphql::*;
 use chrono::{DateTime, Utc};
 use graphql_core::{
-    generic_filters::{DateFilterInput, EqualFilterStringInput, StringFilterInput},
     pagination::PaginationInput,
     standard_graphql_error::{validate_auth, StandardGraphqlError},
     ContextExt,
 };
 use graphql_types::types::*;
-use repository::{
-    location::LocationFilter, DateFilter, EqualFilter, PaginationOption, StockLineFilter,
-    StockLineSort, StockLineSortField, StringFilter,
-};
+use repository::{EqualFilter, PaginationOption, StockLineFilter};
 use service::auth::{Resource, ResourceAccessRequest};
 
 #[derive(Default, Clone)]
 pub struct StockLineQueries;
-
-#[derive(Enum, Copy, Clone, PartialEq, Eq)]
-#[graphql(rename_items = "camelCase")]
-#[graphql(remote = "repository::db_diesel::stock_line::StockLineSortField")]
-pub enum StockLineSortFieldInput {
-    ExpiryDate,
-    NumberOfPacks,
-    ItemCode,
-    ItemName,
-    Batch,
-    PackSize,
-    SupplierName,
-    LocationCode,
-    CostPricePerPack,
-    VvmStatusThenExpiry,
-}
-#[derive(InputObject)]
-pub struct StockLineSortInput {
-    /// Sort query result by `key`
-    key: StockLineSortFieldInput,
-    /// Sort query result is sorted descending or ascending (if not provided the default is
-    /// ascending)
-    desc: Option<bool>,
-}
-
-#[derive(InputObject, Clone)]
-pub struct StockLineFilterInput {
-    pub expiry_date: Option<DateFilterInput>,
-    pub id: Option<EqualFilterStringInput>,
-    pub code: Option<StringFilterInput>,
-    pub name: Option<StringFilterInput>,
-    pub is_available: Option<bool>,
-    pub item_code_or_name: Option<StringFilterInput>,
-    pub search: Option<StringFilterInput>,
-    pub item_id: Option<EqualFilterStringInput>,
-    pub location_id: Option<EqualFilterStringInput>,
-    pub vvm_status_id: Option<EqualFilterStringInput>,
-    pub store_id: Option<EqualFilterStringInput>,
-    pub has_packs_in_store: Option<bool>,
-    pub location: Option<LocationFilterInput>,
-    pub master_list: Option<MasterListFilterInput>,
-    pub is_active: Option<bool>,
-    pub is_program_stock_line: Option<bool>,
-}
-
-impl From<StockLineFilterInput> for StockLineFilter {
-    fn from(f: StockLineFilterInput) -> Self {
-        StockLineFilter {
-            expiry_date: f.expiry_date.map(DateFilter::from),
-            id: f.id.map(EqualFilter::from),
-            code: f.code.map(StringFilter::from),
-            name: f.name.map(StringFilter::from),
-            is_available: f.is_available,
-            item_code_or_name: f.item_code_or_name.map(StringFilterInput::into),
-            search: f.search.map(StringFilterInput::into),
-            item_id: f.item_id.map(EqualFilter::from),
-            location_id: f.location_id.map(EqualFilter::from),
-            store_id: None,
-            vvm_status_id: f.vvm_status_id.map(EqualFilter::from),
-            has_packs_in_store: f.has_packs_in_store,
-            location: f.location.map(LocationFilter::from),
-            master_list: f.master_list.map(|f| f.to_domain()),
-            is_active: f.is_active,
-            is_program_stock_line: f.is_program_stock_line,
-        }
-    }
-}
-
-impl StockLineSortInput {
-    pub fn to_domain(self) -> StockLineSort {
-        StockLineSort {
-            key: StockLineSortField::from(self.key),
-            desc: self.desc,
-        }
-    }
-}
 
 #[Object]
 impl StockLineQueries {
